+++ applications/fishing-map/src/features/map/Map.tsx
@@ -129,6 +129,7 @@ const MapWrapper = (): React.ReactElement => {
         if (hoveredFeatureForDataview) {
           sublayerLegend.currentValue = hoveredFeatureForDataview.value
         }
+        sublayerLegend.currentValue = 32
         return sublayerLegend
       })
     })
@@ -156,7 +157,11 @@ const MapWrapper = (): React.ReactElement => {
         (legend) =>
           document.getElementById(legend.id as string) &&
           createPortal(
-            <MapLegend layer={legend} currentValueClassName={styles.currentValue} />,
+            <MapLegend
+              layer={legend}
+              className={styles.legend}
+              currentValueClassName={styles.currentValue}
+            />,
             document.getElementById(legend.id as string) as Element
           )
       )}