+++ applications/fishing-map/src/features/search/search.slice.ts
@@ -2,14 +2,15 @@ import { createSlice, createAsyncThunk } from '@reduxjs/toolkit'
 import { RootState } from 'store'
 import { AsyncReducerStatus } from 'types'
 import GFWAPI from '@globalfishingwatch/api-client'
-import { Dataset, resolveEndpoint } from '@globalfishingwatch/dataviews-client'
+import { resolveEndpoint } from '@globalfishingwatch/dataviews-client'
+import { Dataset, Vessel, APISearch } from '@globalfishingwatch/api-types'
 
-interface UserState {
+interface SearchState {
   status: AsyncReducerStatus
-  data: any // TODO type search api results
+  data: APISearch<Vessel> | null
 }
 
-const initialState: UserState = {
+const initialState: SearchState = {
   status: AsyncReducerStatus.Idle,
   data: null,
 }