+++ applications/amathea/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @globalfishingwatchapp/amathea
 
+## 0.13.0
+
+### Minor Changes
+
+- db69167: Use types from api-types
+
+### Patch Changes
+
+- Updated dependencies [db69167]
+  - @globalfishingwatch/api-client@1.6.0
+  - @globalfishingwatch/dataviews-client@3.3.0
+  - @globalfishingwatch/react-hooks@3.4.0
+
 ## 0.12.0
 
 ### Minor Changes